+++ CHANGELOG.md
@@ -15,6 +15,7 @@ Added
     - Advanced IM calculation workflow. (--help has more information)
     - Advanced IMs (Steel mf 5story, SAC_steel_MF (3/9 story)
     - Geometric mean calculation for Steel_MF_5_Story
+    - Taghavi and Miranda 2005 calculations added to Advanced IMs 
 
 [19.6.1] - 2019-06-06  
 Added